requirements.txt
# Core ML Libraries
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# Optional ML Libraries (uncomment as needed)
# xgboost>=1.7.0
# lightgbm>=3.3.0
# tensorflow>=2.12.0
# torch>=2.0.0

# Google Cloud Libraries
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-cloud-aiplatform>=1.35.0

# Database Connectivity
pyodbc>=4.0.39

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
joblib>=1.3.0

# Data Visualization (optional)
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Jupyter (for notebooks)
# jupyter>=1.0.0
# ipykernel>=6.23.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
pylint>=2.17.0
